killall(1): Clarify -d, -s and -v options
-d and -v are not equivalent options. The former is more verbose than the
latter and the former does not actually send the signals while the latter does.
Let them have their own paragraphs.

From the point of view of the output, -v is equivalent to -s, so describe them
close to each other. The difference is that former actually sends the signals
and the latter doesn't.

PR:	247411
Approved by:	manpages(0mp)
Differential Revision:	https://reviews.freebsd.org/D25413